starting to pull the analytics events into MendeleyKitiOS (MendeleyKitOSX is to follow)

+++ MendeleyKit/MendeleyKit/Utils/Error Management/MendeleyError.h
@@ -33,7 +33,7 @@
    - 14xx region authentication and networking
    - 15xx General
  */
-typedef NS_ENUM (int, MendeleyErrorCode)
+typedef NS_ENUM (NSInteger, MendeleyErrorCode)
 {
     kMendeleyJSONTypeUnrecognisedErrorCode = 1100,
     kMendeleyJSONTypeNotMappedToModelErrorCode,
